+++ lib/sgchat/go/ext/signal/pkg/libsignalgo/cflags.go
@@ -1,6 +1,6 @@
 package libsignalgo
 
 /*
-#cgo LDFLAGS: -lsignal_ffi -ldl -lm -lz
+#cgo LDFLAGS: -lsignal_ffi -ldl -lm -lz -lstdc++
 */
 import "C"